+++ NEWS.md
@@ -3,6 +3,7 @@
 * progress bar for ui creation in delayed loading module
 * cdisc_dataset and dataset now return R6 class objects (RelationalDataset)
 * Change output of `keys` function to `keys` object. Updated checks in `DatasetConnector` and `dataset`.
+* delayed version of choices_selected
 
 # teal 0.8.3