+++ NEWS.md
@@ -1,8 +1,7 @@
 # teal 0.8.3.9000
 
 * progress bar for ui creation in delayed loading module
-* cdisc_dataset and dataset now return R6 class objects (RelationalDataset)
-* Change output of `keys` function to `keys` object. Updated checks in `DatasetConnector` and `dataset`.
+* Change output of `keys` function to `keys` object. 
 * delayed version of choices_selected
 
 # teal 0.8.3